+++ composer-c1x/src/main/java/io/piano/android/composer/c1x/ShowRecommendationsController.kt
@@ -32,6 +32,7 @@ class ShowRecommendationsController(event: Event<ShowRecommendations>) :
             jsInterface: WidgetJs,
             dialogFragment: ShowRecommendationsDialogFragment? = null
         ) {
+            settings.domStorageEnabled = true
             settings.javaScriptEnabled = true
             jsInterface.init(dialogFragment, this)
             addJavascriptInterface(jsInterface, JAVASCRIPT_INTERFACE)