+++ id/id/src/main/java/io/piano/android/id/models/PianoIdApi.kt
@@ -2,12 +2,14 @@ package io.piano.android.id.models
 
 import okhttp3.ResponseBody
 import retrofit2.Call
+import retrofit2.http.Body
 import retrofit2.http.Field
 import retrofit2.http.FieldMap
 import retrofit2.http.FormUrlEncoded
 import retrofit2.http.GET
 import retrofit2.http.Header
 import retrofit2.http.POST
+import retrofit2.http.PUT
 import retrofit2.http.Query
 import retrofit2.http.Url
 
@@ -39,6 +41,14 @@ internal interface PianoIdApi {
         @Query("client_id") aid: String,
         @Header("Authorization") accessToken: String,
         @Query("form_name") formName: String?
+    ): Call<PianoUserProfile>
+
+    @PUT
+    fun putUserInfo(
+        @Url url: String,
+        @Query("aid") aid: String,
+        @Header("Authorization") accessToken: String,
+        @Body updateRequest: ProfileUpdateRequest
     ): Call<PianoUserProfile>
 
     @FormUrlEncoded