+++ id/id/src/main/java/io/piano/android/id/models/ProfileUpdateRequest.kt
@@ -1,0 +1,12 @@
+package io.piano.android.id.models
+
+import com.squareup.moshi.Json
+import com.squareup.moshi.JsonClass
+
+@JsonClass(generateAdapter = true)
+class ProfileUpdateRequest(
+    @Json(name = "form_name")
+    val formName: String,
+    @Json(name = "custom_field_values")
+    val customFields: List<CustomField>
+)